stripe_backend/requirements.txt
asgiref==3.8.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-cors-headers==4.3.1
djangorestframework==3.12.4
idna==3.3
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
stripe
typing_extensions==4.11.0
urllib3==1.26.7
